# openapi/climate-hub-v1.yaml
openapi: "3.0.3"
info:
  title: Climate Hub API — Need Engine Closed Loop
  version: "1.1.0"
  description: |
    REST API локальной платформы управления микроклиматом здания.
    Реализован замкнутый контур Need Engine: Environment → Need → Capability →
    Device → Command → Effect → Satisfaction.
servers:
  - url: /api/v1
    description: Climate Hub API
paths:
...
  /rooms/{roomId}/environment:
    get:
      operationId: getRoomEnvironment
      summary: Текущее состояние среды помещения
      tags: [Environment]
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Текущее состояние
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomEnvironmentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /rooms/{roomId}/environment/history:
    get:
      operationId: getRoomEnvironmentHistory
      summary: История параметров среды
      tags: [Environment]
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: parameter
          in: query
          schema:
            type: string
            enum: [temperature, humidity, co2]
      responses:
        "200":
          description: История параметра
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentHistoryResponse"

  # ────────────────────── Devices ──────────────────────
...
components:
  schemas:
...
    MeasurementQuality:
      type: string
      enum: [valid, estimated, stale, outOfRange, sensorError, rejected, unknown]
    RoomEnvironmentResponse:
      type: object
      required:
        - roomId
        - roomName
        - status
        - parameters
        - updatedAt
      properties:
        roomId:
          type: string
          format: uuid
        roomName:
          type: string
        status:
          type: string
          enum: [normal, warning, critical, offline, stale, unknown]
        parameters:
          type: object
          properties:
            temperature:
              $ref: "#/components/schemas/EnvironmentParameter"
            relativeHumidity:
              $ref: "#/components/schemas/EnvironmentParameter"
            co2:
              $ref: "#/components/schemas/EnvironmentParameter"
        updatedAt:
          type: string
          format: date-time
    EnvironmentParameter:
      type: object
      required:
        - value
        - unit
        - measuredAt
        - receivedAt
        - quality
        - sourceDeviceId
      properties:
        value:
          type: number
        unit:
          type: string
        measuredAt:
          type: string
          format: date-time
        receivedAt:
          type: string
          format: date-time
        quality:
          $ref: "#/components/schemas/MeasurementQuality"
        sourceDeviceId:
          type: string
          format: uuid
        target:
          $ref: "#/components/schemas/EnvironmentTarget"
    EnvironmentTarget:
      type: object
      properties:
        minimum:
          type: number
        maximum:
          type: number
        preferred:
          type: number
    EnvironmentHistoryResponse:
      type: object
      required:
        - roomId
        - parameter
        - unit
        - from
        - to
        - aggregation
        - points
      properties:
        roomId:
          type: string
          format: uuid
        parameter:
          type: string
          enum: [temperature, humidity, co2]
        unit:
          type: string
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        aggregation:
          type: string
          enum: [raw, 1m, 15m, 1h]
        points:
          type: array
          items:
            $ref: "#/components/schemas/EnvironmentHistoryPoint"
    EnvironmentHistoryPoint:
      type: object
      required:
        - timestamp
        - value
        - quality
      properties:
        timestamp:
          type: string
          format: date-time
        value:
          type: number
        quality:
          $ref: "#/components/schemas/MeasurementQuality"
...
    ApiProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errorCode:
          type: string
        traceId:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
  responses:
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiProblemDetails"
...